requirements.txt
PyQt6==6.11.0
PyQt6-Qt6==6.11.0
PyQt6-sip==13.11.1
google-generativeai==0.5.4
edge-tts==6.1.9
SpeechRecognition==3.10.1
pyaudio==0.2.14
pywin32==306
python-dotenv==1.0.1
numpy==1.26.4
soundfile==0.12.1
mss==9.0.1
Pillow==10.3.0
qasync==0.27.1
